Fix WorksspaceService exprot
